requirements.txt
numpy==1.24.3
pandas==2.0.1
matplotlib==3.7.1
plotly==5.14.1
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
pydantic==1.10.8
tqdm==4.65.0
beautifulsoup4==4.12.2
lxml==4.9.3
dash==2.14.2
dash-bootstrap-components==1.5.0
slack-sdk==3.26.1
tenacity==8.2.3
prometheus-client==0.19.0 